requirements.txt
django==5.1.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
pillow
requests
psycopg2-binary
python-dotenv
celery
redis
django-redis
django-celery-results
django-celery-beat
channels
daphne
channels-redis
gunicorn
whitenoise
dj-database-url
